+++ js/large_image/large_image.js
@@ -73,8 +73,13 @@ jQuery(document).ready(function() {
         updateAnnotation(annotation);
     });
 
-    anno.addHandler("onAnnotationRemoved", function(annotation) {
+    anno.addHandler("beforeAnnotationRemoved", function(annotation) {
+        var isConfirmed = confirm("This annotation will be deleted. This action cannot be reversed. Are you sure?");
+        if (isConfirmed) {
             deleteAnnotation(annotation);
+        } else {
+            return false;
+        }
     });
 
     executeCommonLoadOperations();
